requirements.txt
pycryptodome
deflate
termcolor
pypsexec
